requirements.txt
streamlit>=1.32.0
torch>=2.0.0
transformers>=4.38.0
accelerate>=0.26.0
Pillow>=10.0.0
urllib3>=2.0.0
